pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nonebot-plugin-nodejsphira"
version = "0.1.3"
description = "Phira Server Manager plugin for NoneBot2"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "chuzouX" },
]
dependencies = [
    "nonebot2>=2.4.4",
    "nonebot-adapter-qq>=1.6.6",
    "httpx>=0.28.1",
    "cryptography>=46.0.4",
    "nonebot-plugin-htmlrender>=0.3.6",
]

[project.urls]
Homepage = "https://github.com/chuzouX/nonebot_plugin_nodejsphira"
Repository = "https://github.com/chuzouX/nonebot_plugin_nodejsphira"

[tool.hatch.build.targets.sdist]
include = ["/nonebot_plugin_nodejsphira"]

[tool.hatch.build.targets.wheel]
packages = ["nonebot_plugin_nodejsphira"]